+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 -   Glow Berries are now a source for yellow dye [\#196](https://github.com/EnigmaticaModpacks/Enigmatica10/pull/196)
 -   The Market has dropped their prices once again! Rejoice! [\#196](https://github.com/EnigmaticaModpacks/Enigmatica10/pull/196)
 -   Ronaza's Contact no longer uses Sculk Catalyst on account of it eating the ritual. Nom. [\#202](https://github.com/EnigmaticaModpacks/Enigmatica10/pull/202)
+-   Add missing blocks to Eldritch Miner [\#204](https://github.com/EnigmaticaModpacks/Enigmatica10/pull/204)
 
 #### 🦟 Bugs Fixed